# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "dealerService implements dealer micro service\n\nThe purpose of this application is to provides api's to perform CURD operations on dealer object.\nCurrently only get endpoints are available.\ndealerService is divided into 4 file.\n1. dealerRoutes.go  -\u003e contain routes.\n2. dealerHandler.go -\u003e containing handler functions.\n3. dealerModel.go   -\u003e containing models.\n4. dealerUtils.go   -\u003e containing util functions.",
    "title": "implements dealer micro service\ntitle: Dealer micro service",
    "contact": {
      "name": "Qasim Hasnain",
      "email": "mqhasnain@tekion.com"
    },
    "version": "1.0.0"
  },
  "basePath": "/tdealer",
  "paths": {
    "/contact/{cid}": {
      "get": {
        "description": "By default /contact/{cid} returns complete dealer contact object.\nIn case you need only certain fields, you can specify an optional query parameter \"fields\",\npassing a list of comma separated fields you want in response.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "dealerContact"
        ],
        "summary": "Returns dealer contact identified by dealer contact id passed as part of url",
        "operationId": "readDealerContact",
        "parameters": [
          {
            "type": "string",
            "description": "unique identifier of the dealer contact",
            "name": "cid",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "unique identifier of the dealer",
            "name": "dealerid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "client type",
            "name": "clientid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "current tenant name",
            "name": "tenantname",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "auth token",
            "name": "tekion-api-token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "e.g /contact/{cid}?fields=user,userDisplayName,userDisplayTitle",
            "name": "fields",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "dealer contact object",
            "schema": {
              "$ref": "#/definitions/dealerContact"
            }
          },
          "204": {
            "description": "dealer contact not found in data base"
          },
          "400": {
            "description": "error querying data base"
          }
        }
      }
    },
    "/contacts": {
      "get": {
        "description": "By default /contacts returns list of complete dealer contacts objects.\nIn case you need only certain fields, you can specify an optional query parameter \"fields\",\npassing a list of comma separated fields you want in response.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "dealerContacts"
        ],
        "summary": "Returns list of dealer contacts identified by dealer id passed in header",
        "operationId": "readDealerContacts",
        "parameters": [
          {
            "type": "string",
            "description": "unique identifier of the dealer",
            "name": "dealerid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "client type",
            "name": "clientid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "current tenant name",
            "name": "tenantname",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "auth token",
            "name": "tekion-api-token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "e.g /contacts?fields=user,userDisplayName,userDisplayTitle",
            "name": "fields",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "list of dealer contacts",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/dealerContact"
              }
            }
          },
          "204": {
            "description": "dealer contacts not found in data base"
          },
          "400": {
            "description": "error querying data base"
          }
        }
      }
    },
    "/dealer": {
      "get": {
        "description": "By default /dealer returns complete dealer object.\nIn case you need only certain fields, you can specify an optional query parameter \"fields\",\npassing a list of comma separated fields you want in response.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "dealer"
        ],
        "summary": "Returns Dealer identified by the dealer id",
        "operationId": "readDealer",
        "parameters": [
          {
            "type": "string",
            "description": "unique identifier of the dealer",
            "name": "dealerid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "client type",
            "name": "clientid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "current tenant name",
            "name": "tenantname",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "auth token",
            "name": "tekion-api-token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "e.g /dealer?fields=dealerDoingBusinessAsName,vehicleDamage,dealerAddress",
            "name": "fields",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "dealer object",
            "schema": {
              "$ref": "#/definitions/dealer"
            }
          },
          "204": {
            "description": "dealer not found in data base"
          },
          "400": {
            "description": "error querying data base"
          }
        }
      },
      "patch": {
        "description": "By default /dealerDtls returns complete dealer details.\nIn case you need only certain fields, you can specify an optional query parameter \"fields\",\npassing a list of comma separated fields you want in response.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "dealer"
        ],
        "summary": "Returns dealer list of columns to update",
        "operationId": "patchDealer",
        "parameters": [
          {
            "type": "string",
            "description": "unique identifier of the dealer",
            "name": "dealerid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "client type",
            "name": "clientid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "current tenant name",
            "name": "tenantname",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "auth token",
            "name": "tekion-api-token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "e.g /dealer?fields=dealerDoingBusinessAsName,vehicleDamage,dealerAddress",
            "name": "fields",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "dealer object",
            "schema": {
              "$ref": "#/definitions/dealer"
            }
          },
          "204": {
            "description": "dealer not found in data base"
          },
          "400": {
            "description": "error querying data base"
          }
        }
      }
    },
    "/dealers": {
      "post": {
        "description": "By default /lstDealer returns complete dealer list.\nIn case you need only certain fields, you can specify an optional query parameter \"fields\",\npassing a list of comma separated fields you want in response.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "listDealersReq"
        ],
        "summary": "Returns Dealer list",
        "operationId": "dealerList",
        "parameters": [
          {
            "type": "string",
            "description": "unique identifier of the dealer",
            "name": "dealerid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "client type",
            "name": "clientid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "current tenant name",
            "name": "tenantname",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "auth token",
            "name": "tekion-api-token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "e.g /dealer?fields=dealerDoingBusinessAsName,vehicleDamage,dealerAddress",
            "name": "fields",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "dealer object",
            "schema": {
              "$ref": "#/definitions/dealer"
            }
          },
          "204": {
            "description": "dealer not found in data base"
          },
          "400": {
            "description": "error querying data base"
          }
        }
      }
    },
    "/fixedoperation": {
      "get": {
        "description": "By default /fixedoperation returns list of complete fixed operation objects.\nIn case you need only certain fields, you can specify an optional query parameter \"fields\",\npassing a list of comma separated fields you want in response.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "fixedOperation"
        ],
        "summary": "Returns list of fixed operations identified by dealer id passed in header",
        "operationId": "readFixedOperation",
        "parameters": [
          {
            "type": "string",
            "description": "unique identifier of the dealer",
            "name": "dealerid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "client type",
            "name": "clientid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "current tenant name",
            "name": "tenantname",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "auth token",
            "name": "tekion-api-token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "e.g /fixedoperations?field=serviceAdvisors,floorCapacity,appointmentHour,appointmentCapacity",
            "name": "fields",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "list of fixed operations",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/fixedOperation"
              }
            }
          },
          "204": {
            "description": "fixed operations not found in data base"
          },
          "400": {
            "description": "error querying data base"
          }
        }
      }
    },
    "/goal/{gid}": {
      "get": {
        "description": "By default /goal/{gid} returns complete dealer goal object.\nIn case you need only certain fields, you can specify an optional query parameter \"fields\",\npassing a list of comma separated fields you want in response.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "dealerGoal"
        ],
        "summary": "Returns dealer goal identified by dealer goal id passed as part of url",
        "operationId": "readDealerGoal",
        "parameters": [
          {
            "type": "string",
            "description": "unique identifier of the dealer goal",
            "name": "gid",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "unique identifier of the dealer",
            "name": "dealerid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "client type",
            "name": "clientid",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "current tenant name",
            "name": "tenantname",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "auth token",
            "name": "tekion-api-token",
            "in": "header",
            "required": true
          },
          {
            "type": "string",
            "description": "e.g /goal/{id}?fields=hoursPerRepairOrderAdvisorGoal,totalHoursAdvisorGoal,averageLaborRateAdvisorGoal",
            "name": "fields",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "dealer goal object",
            "schema": {
              "$ref": "#/definitions/dealerGoal"
            }
          },
          "204": {
            "description": "dealer goal not found in data base"
          },
          "400": {
            "description": "error querying data base"
          }
        }
      }
    },
...
  },
...
}